requirements.txt
asgiref==3.8.1
typing_extensions==4.9.0
beautifulsoup4==4.12.2
bs4==0.0.2
Django
django-cors-headers==4.4.0
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
requests==2.31.0
requests-oauthlib==2.0.0
selenium==4.23.1
sqlparse==0.5.0
redis==5.0.8
celery==5.4.0
django-celery-beat==2.6.0